机器人/None_QQ/test/pyproject.toml
[tool.poetry]
name = "None_QQ"
version = "0.1.0"
description = "None_QQ"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ['nonebot_plugin_biliav', 'nonebot_plugin_simplemusic', 'nonebot_plugin_minesweeper', 'nonebot_plugin_charpic', 'nonebot_plugin_code', 'nonebot_plugin_remake', 'nonebot_plugin_petpet', 'nonebot_plugin_logo']
plugin_dirs = ["none_qq/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
